Adding "align-items" (need testing)
Depiste this update it need testing, i have no idea how responsive is the joystream, so it's possibly the responsivity. But i will need it being tested with different resolutions. I can't test it wihout having a bigger screen. SO you guys could test it for me, i don't know if it will work